pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cortex"
version = "0.1.0"
description = "Tiered memory system for AI assistants with emotional decay and consolidation"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
    { name = "Josh Heidorn" }
]
keywords = ["memory", "ai", "assistant", "redis", "postgres", "embeddings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "redis>=5.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "apscheduler>=3.10.0",
    "structlog>=24.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "mem0ai>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
local-embeddings = [
    "sentence-transformers>=2.5.0",
]

[project.scripts]
cortex = "cortex.cli:app"

[project.urls]
Homepage = "https://github.com/jheidorn/cortex"
Documentation = "https://github.com/jheidorn/cortex#readme"
Issues = "https://github.com/jheidorn/cortex/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/cortex"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
strict = true
